Added Jupyter Notebook
In this project, I build a XGBoost classification model, using the Invistico Airline dataset that I previously used in building the decision tree and random forest models, and then compare the performance of the three models. The ultimate goal is to decide which model is best.

Lastly, I’ll explore the features importance that most contribute to customer satisfaction, as revealed by my XGBoost model.